fix reopen notify
executorIdがnullになるケースを考慮

+++ src/handleReopenNotify.js
@@ -17,19 +17,23 @@ import { ButtonStyle, ComponentType } from 'discord.js'
  */
 export async function handleReopenNotify(logger, entry, thread, setting) {
   logger.info(`"${thread.name}" (${thread.id}) has been reopened.`)
-  const message = await thread.send({
-    content: setting.onReopen(entry.executorId),
-    components,
-  })
-  const interaction = await message
-    .awaitMessageComponent({
-      filter: i => i.user.id === entry.executorId,
-      time: 1000 * 60 * 5,
+  if (entry.executorId) {
+    const message = await thread.send({
+      content: setting.onReopen(entry.executorId),
+      components,
     })
-    .catch(() => null)
-  if (interaction) {
-    await message.delete()
-    await thread.setArchived()
+    const interaction = await message
+      .awaitMessageComponent({
+        filter: i => i.user.id === entry.executorId,
+        time: 1000 * 60 * 5,
+      })
+      .catch(() => null)
+    if (interaction) {
+      await message.delete()
+      await thread.setArchived()
+    }
+  } else {
+    await thread.send(setting.onReopen())
   }
 }
